test existence and uniqueness in db
